Changed file structure and added quiz landing page
