+++ src/app/quiz/mcq/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+import { useState } from 'react';
+import Mcq from '../components/Mcq';
+import '../styles.css';
+
+export default function multiplechoice() {
+    return <Mcq></Mcq>;
+}
