+++ src/app/quiz/scenario-game/page.tsx
@@ -1,0 +1,5 @@
+'use client';
+import { useState } from 'react';
+import '../styles.css';
+
+export default function truefalse() {}
